Add org-fragtog-preview-delay (#25)
* Add org-fragtog-preview-delay

Enable delayed preview so that moving through LaTeX fragments
does not toggle them unless the cursor stops and Emacs is idle.

* Don't use timers if delay is 0

* Renew cursor frag upon timed disable

The Emacs command loop adjusts point to skip invisible
parts of the buffer /after/ all functions in `post-command-hook'
are done. This causes `org-fragtog--post-cmd' to obtain
an "incorrect" (non-adjusted) point for inspection with
`org-fragtog--cursor-frag'. When a timed `org-fragtog--disable-frag'
runs, make sure it obtains the adjusted point by running
`org-fragtog--cursor-frag' again.

* Don't enable if not disabled

Don't enable fragments when the cursor left before
a timed disable occured.

* Enable based on overlay

If previous fragment did not have an `org-latex-overlay',
enable it regardless of the timer. This makes sure that
new fragments and edited fragments are enabled when expected.

* Expire fragment on renewed disable only

If org-fragtog runs without a delay, no timers are created.
It is not necessary to expire `org-fragtog--timer' in
`org-fragtog--enable-frag' in this case.